% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucular.R
\name{adjust_X_chr}
\alias{adjust_X_chr}
\title{adjust_X_chr}
\usage{
adjust_X_chr(ff_list, X_chr, Xfactor)
}
\arguments{
\item{ff_list}{list of ff objects each representing data of one chromosome. IMPORTANT: the data in this list will be changed!}

\item{X_chr}{(character) name of the X chromosome in \code{ff_list}}

\item{Xfactor}{(numeric) factor by which the counts should be adjusted (2 for male cell lines, 4/3 for male/female mixtures, i.e. embryo's or unspecific adult flies)}
}
\value{
reference to the ff_list, which is now adjusted. IMPORTANT: this is the same ff_list as the input, no copy is created!
}
\description{
Adjusts lower X chromosome counts due to cells being male or a mixture of male/female.
}
