% Generated by roxygen2 (4.0.1): do not edit by hand
\name{center_list}
\alias{center_list}
\title{center_list}
\usage{
center_list(x)
}
\arguments{
\item{x}{list of numeric vectors}
}
\value{
centered x (mean = 0)
}
\description{
centers all list elements by their combined mean
}
\author{
Mark Heron
}

