% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_list_list_functions.R
\name{compare_ff_list_list}
\alias{compare_ff_list_list}
\title{compare_ff_list_list}
\usage{
compare_ff_list_list(data_list, comparison_function)
}
\arguments{
\item{data_list}{list of the ff_list objects to compare}

\item{comparison_function}{comparison function to use, e.g. cor_ff_list}
}
\value{
comparison matrix
}
\description{
Applies a comparison function pairwise between all data_list elements and saves the results in a matrix
}
