% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_list_list_functions.R
\name{compare_ff_list_list_vs_ff_list_list}
\alias{compare_ff_list_list_vs_ff_list_list}
\title{compare_ff_list_list_vs_ff_list_list}
\usage{
compare_ff_list_list_vs_ff_list_list(first_list, second_list,
  comparison_function)
}
\arguments{
\item{first_list}{of the ff_list objects to compute the correlations from.}

\item{second_list}{of the ff_list objects to compute the correlations to.}

\item{comparison_function}{comparison function to use, e.g. cor_ff_list}
}
\value{
comparison matrix
}
\description{
Applies a comparison function pairwise between the list elements of two different lists and saves the results in a matrix
}
\author{
Mark Heron
}

