% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nucular.R
\name{convertRaw2Rdata}
\alias{convertRaw2Rdata}
\title{convertRaw2Rdata}
\usage{
convertRaw2Rdata(name, folder = "", extension = ".tabular")
}
\arguments{
\item{name}{of the data file}

\item{folder}{in which the file is to be found}

\item{extension}{of the data file, standard is ".tabular"}
}
\description{
Converts raw data tables from the galaxy pipeline to lists of matricies that is saved as an Rdata for futher use.
}
\author{
Mark Heron
}

