% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_list_list_functions.R
\name{convertSparse2occ_ff_list_list}
\alias{convertSparse2occ_ff_list_list}
\title{convertSparse2occ_ff_list_list}
\usage{
convertSparse2occ_ff_list_list(data_list, lengths)
}
\arguments{
\item{data_list}{list of different sparse representation matrix lists}

\item{lengths}{list of chromosome lengths, \code{names} must match the \code{names} of data_list sub-elements (i.e. \code{names(data_list[[1]])})}
}
\value{
list of lists of ff vectors with genomic occuopancy
}
\description{
Individually converts a list of different sparse representation of dyad positions to nucleosome occupancy as list of ff vectors.
}

