% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cor_ff_list}
\alias{cor_ff_list}
\title{cor_ff_list}
\usage{
cor_ff_list(x, y)
}
\arguments{
\item{x}{a list of ff vectors}

\item{y}{a second list of ff vectors}
}
\value{
pearson correlation between x and y
}
\description{
Computes the correlation between two vectors.
}
\author{
Mark Heron
}

