% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ff_list_list_functions.R
\name{cor_ff_list_list_vs_ff_list_list}
\alias{cor_ff_list_list_vs_ff_list_list}
\title{cor_ff_list_list_vs_ff_list_list}
\usage{
cor_ff_list_list_vs_ff_list_list(first_list, second_list)
}
\arguments{
\item{first_list}{of the ff_list objects to compute the correlations from.}

\item{second_list}{of the ff_list objects to compute the correlations to.}
}
\value{
correlation matrix
}
\description{
Computes the pairwise correlations between the list elements of two different lists and saves them in a matrix
}
\author{
Mark Heron
}

