% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nucular.R
\name{cor_nucs}
\alias{cor_nucs}
\title{cor_nucs}
\usage{
cor_nucs(data_list1, data_list2, lengths)
}
\arguments{
\item{data_list1}{first ff_list of nucleosome positions}

\item{data_list2}{second ff_list of nucleosome positions}

\item{lengths}{list of chromosome lengths (names must match subset of the data_list's )}
}
\value{
correlation pearson correlation between the nucleosome occupancies
}
\description{
Calculates the pearson correlation between two nucleosome occupancy profiles given their sparse representation.
}
\author{
Mark Heron
}

