% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucular.R
\name{cor_nucs_multiple}
\alias{cor_nucs_multiple}
\title{cor_nucs_multiple}
\usage{
cor_nucs_multiple(data_list, lengths)
}
\arguments{
\item{data_list}{list of ff_list of nucleosome dyad positions}

\item{lengths}{list of chromosome lengths (\code{names} must match subset of the data_list's \code{names})}
}
\value{
top triangle matrix of the pearson correlations between nucleosome occupancies
}
\description{
Computes all pairwise pearson correlations between the nucleosome occupancies.
}

