% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucular.R
\name{cor_nucs_using_single_vecs}
\alias{cor_nucs_using_single_vecs}
\title{cor_nucs_using_single_vecs}
\usage{
cor_nucs_using_single_vecs(data_list1, data_list2, lengths)
}
\arguments{
\item{data_list1}{first ff_list of nucleosome positions}

\item{data_list2}{second ff_list of nucleosome positions}

\item{lengths}{list of chromosome lengths (\code{names} must match subset of the data_list's \code{names})}
}
\value{
correlation pearson correlation between the nucleosome occupancies
}
\description{
Calculates the pearson correlation between two nucleosome occupancy profiles given their sparse representation.
Same as cor_nucs but uses an older non-optimized method, being kept for testing purposis.
}
