% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cov_ff}
\alias{cov_ff}
\title{cov_ff}
\usage{
cov_ff(x, y, centered = FALSE)
}
\arguments{
\item{x}{a ff vector}

\item{y}{a second ff vector}

\item{centered}{boolean, is the mean already = 0 ?}
}
\value{
covariance between x and y
}
\description{
Computest the covariance between two ff vectors.
}
\author{
Mark Heron
}

