% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ff_list_functions.R
\name{df_list_to_ffdf_list}
\alias{df_list_to_ffdf_list}
\title{df_list_to_ffdf_list}
\usage{
df_list_to_ffdf_list(df_list)
}
\arguments{
\item{df_list}{to convert to a ffdf_list}
}
\value{
list of ffdf objects
}
\description{
Creates a list where each element of df_list is saved in as an ffdf object
}
\author{
Mark Heron
}

