% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ff_functions.R
\name{ff_functions}
\alias{ff_functions}
\title{function variations to run on ff vectors}
\description{
Copies of several functions that are originally for numeric vectors, but in this *_ff variations work on ff vectors
}
\author{
Mark Heron
}

