% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_list_functions.R
\name{ff_list_functions}
\alias{ff_list_functions}
\title{function variations to run on ff vector lists}
\description{
Copies of several functions that are originally for numeric vectors, but in this *_ff_list variations work on lists of ff vectors
}
\author{
Mark Heron
}

