% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ff_list_list_functions.R
\name{ff_list_list_functions}
\alias{ff_list_list_functions}
\title{function variations to run on lists of vectors, handling them as one vector}
\description{
Copies of several functions that are originally for numeric vectors, but in this _list variations work on list of vectors
}
\author{
Mark Heron
}

