% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_list_functions.R
\name{hist_ff_list}
\alias{hist_ff_list}
\title{hist_ff_list}
\usage{
hist_ff_list(ff_list, column, ...)
}
\arguments{
\item{ff_list}{list of ff matricies}

\item{column}{of the ff_list elements to plot the histogram for}

\item{...}{parameters to be passed on to \code{\link{hist}}}
}
\description{
Plots a histogram for the values in \code{column} for every ff_list element.
}
