% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ff_list_functions.R
\name{intra_scaling_ff_list}
\alias{intra_scaling_ff_list}
\title{intra_scaling_ff_list}
\usage{
intra_scaling_ff_list(ff_list)
}
\arguments{
\item{ff_list}{where each vector should be scaled individually}
}
\value{
intra scaled ff_list
}
\description{
Scales each ff_vector in the list to average 1.
Simple wrapper for lapply '/' mean
}

