% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_list_functions.R
\name{likelihood_ff_list}
\alias{likelihood_ff_list}
\title{likelihood_ff_list}
\usage{
likelihood_ff_list(predictions, measurements)
}
\arguments{
\item{predictions}{(list of ff vectors)}

\item{measurements}{(list of ff vectors)}
}
\value{
(numeric) likelihood
}
\description{
Computes a simple likelihood of a list of prediction vectors given a matching list of measurement vectors.
}
