% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ff_list_list_functions.R
\name{likelihood_ff_list_list_vs_ff_list_list}
\alias{likelihood_ff_list_list_vs_ff_list_list}
\title{likelihood_ff_list_list_vs_ff_list_list}
\usage{
likelihood_ff_list_list_vs_ff_list_list(first_list, second_list)
}
\arguments{
\item{first_list}{of the ff_list objects}

\item{second_list}{of the ff_list objects}
}
\value{
likelihood matrix
}
\description{
Computes the pairwise likelihood between the list elements of two different lists and saves them in a matrix
}
\author{
Mark Heron
}

