% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list_functions}
\alias{list_functions}
\title{function variations to run on lists of vectors, handling them as one vector}
\description{
Copies of several functions that are originally for numeric vectors, but in this _list variations work on list of vectors
}
\author{
Mark Heron
}

