% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_list_functions.R
\name{make_prob_ff_list}
\alias{make_prob_ff_list}
\title{center_ff_list}
\usage{
make_prob_ff_list(x)
}
\arguments{
\item{x}{(list of ff_vectors)}
}
\value{
(list of ff_vectors) probability version of x (sum == 1)
}
\description{
use center_list
scale_ff_list
}
\details{
use scale_list
intra_scaling_ff_list

use intre_scaling_list
make_prob_ff_list

Scales all list elements so their sum is 1, i.e. if there are no negative values they can be used as probabilities.
}
