% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucular.R
\name{mask_Ns_in_ff_list}
\alias{mask_Ns_in_ff_list}
\title{mask_Ns_in_ff_list}
\usage{
mask_Ns_in_ff_list(ff_list, genome_fasta)
}
\arguments{
\item{ff_list}{to be masked}

\item{genome_fasta}{(DNAStringSet) of the genome}
}
\value{
masked ff_list (cloned so the original is \emph{not} changed)
}
\description{
Masks an ff_list based on N's in the genomic sequence.
}

