% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nucular.R
\name{mask_repeats}
\alias{mask_repeats}
\title{mask_repeats}
\usage{
mask_repeats(occ_single, mask_file_folder, repeats_to_mask,
  nuc_half_width = 73)
}
\arguments{
\item{occ_single}{(list of ff vectors) each ff representing one chromosome}

\item{mask_file_folder}{(character) name of the folder where the repeatMasker files for each chromosome can be found}

\item{repeats_to_mask}{(character vector) of types of repeats to mask}

\item{nuc_half_width}{(integer) amount to additionally mask to the left and right of the repeat}
}
\value{
the masked occupancy ff vector list
}
\description{
Masks repeat regions read in from repeatMasker output files.
}
\author{
Mark Heron
}

