% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucular.R
\docType{package}
\name{nucular-package}
\alias{nucular-package}
\title{nucleosomal utilities for lists}
\description{
Provides convienience functions for handling (nucleosomal) genomic data as lists of ff vectors (or vectors).
The focus lies on comparing the genomic data by computing correlations and other comparison scores between the vector lists.
}
\details{
(nuccontrol and nucppp actually do reporting)
}
\author{
Mark Heron
}
