% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_list_functions.R
\name{rmse_ff_list}
\alias{rmse_ff_list}
\title{rmse_ff_list}
\usage{
rmse_ff_list(predictions, measurements)
}
\arguments{
\item{predictions}{a list of ff vectors}

\item{measurements}{a second list of ff vectors}
}
\value{
root mean square error
}
\description{
Computes the root mean square error between two ff_lists.
}

