% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_functions.R
\name{scale_list}
\alias{scale_list}
\title{scale_list}
\usage{
scale_list(x)
}
\arguments{
\item{x}{list of numeric vectors}
}
\value{
scaled x (mean = 1)
}
\description{
scales all list elements by their combined mean
}
\author{
Mark Heron
}

