% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{sum_list}
\alias{sum_list}
\title{sum_list}
\usage{
sum_list(x, na.rm = TRUE)
}
\arguments{
\item{x}{list of numeric vectors}

\item{na.rm}{should NA values be ignored (see \code{\link{sum}})}
}
\value{
(numeric) sum of all values in the list
}
\description{
Calculates the sum of all list elements together.
}
