% Generated by roxygen2 (4.0.1): do not edit by hand
\name{uniquify_ff_list}
\alias{uniquify_ff_list}
\title{uniquify_ff_list}
\usage{
uniquify_ff_list(ff_list)
}
\arguments{
\item{ff_list}{list of ff matrices each representing data of one chromosome}
}
\value{
demultiplexed
}
\description{
Removes duplicate fragments in a list of ff objects representing all reads for a chromosome.
}
\author{
Mark Heron
}

