% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucular.R
\name{adjust_X_chr}
\alias{adjust_X_chr}
\title{adjust_X_chr}
\usage{
adjust_X_chr(ff_list, X_chr, Xfactor)
}
\arguments{
\item{ff_list}{list of ff objects each representing data of one chromosome. IMPORTANT: the data in this list will be changed!}

\item{X_chr}{name of the X chromosome in ff_list}

\item{Xfactor}{factor by which the counts should be adjusted (2 for male cell lines, 4/3 for male/female mixtures (embryo's or unspecific adult flies))}
}
\value{
adjusted ff_list
}
\description{
Adjusts lower X chromosome counts due cells being male or a mixture of male/female.
}
\author{
Mark Heron
}

