% Generated by roxygen2 (4.0.1): do not edit by hand
\name{array_list_to_ff_list}
\alias{array_list_to_ff_list}
\title{array_list_to_ff_list}
\usage{
array_list_to_ff_list(array_list)
}
\arguments{
\item{array_list}{to convert to a ff_list}
}
\value{
list of ff objects
}
\description{
Creates a list where each element of array_list is saved in as an ff object
}
\author{
Mark Heron
}

