% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_list_functions.R
\name{array_list_to_ff_list}
\alias{array_list_to_ff_list}
\title{array_list_to_ff_list}
\usage{
array_list_to_ff_list(array_list)
}
\arguments{
\item{array_list}{to convert to a ff_list}
}
\value{
list of ff objects
}
\description{
Creates a copy of array_list where each element is saved as an ff object.
}

