% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ff_list_functions.R
\name{center_ff_list}
\alias{center_ff_list}
\title{center_ff_list}
\usage{
center_ff_list(x)
}
\arguments{
\item{x}{list of ff_vectors}
}
\value{
centered x (mean = 0)
}
\description{
centers all list elements by their combined mean
}
\author{
Mark Heron
}

