% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{center_list}
\alias{center_list}
\title{center_list}
\usage{
center_list(x)
}
\arguments{
\item{x}{list of numeric vectors (or ff vectors)}
}
\value{
(list of numeric (ff) vectors) centered \code{x} (i.e., mean_list == 0)
}
\description{
Centers all list elements by their combined mean.
Works with lists of ff vectors.
}
