% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{center_list}
\alias{center_list}
\title{center_list}
\usage{
center_list(x)
}
\arguments{
\item{x}{list of numeric vectors}
}
\value{
centered x (mean = 0)
}
\description{
centers all list elements by their combined mean
works with ff_lists
}
\author{
Mark Heron
}

