% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_list_functions.R
\name{convertSparse2occ_ff_list}
\alias{convertSparse2occ_ff_list}
\title{convertSparse2occ_ff_list}
\usage{
convertSparse2occ_ff_list(sparse, lengths)
}
\arguments{
\item{sparse}{list of sparse representation matricies}

\item{lengths}{list of chromosome lengths, names must match those of sparse}
}
\value{
list of ff vectors with genomic occuopancy
}
\description{
Converts sparse representation of dyad positions to nucleosome occupancy as list of ff vectors.
}
\author{
Mark Heron
}

