% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convertSparse2occ_ff_list_list}
\alias{convertSparse2occ_ff_list_list}
\title{convertSparse2occ_ff_list_list}
\usage{
convertSparse2occ_ff_list_list(data_list, lengths)
}
\arguments{
\item{data_list}{list of different sparse representation matrix lists}

\item{lengths}{list of chromosome lengths, names must match those of data_list subelements}
}
\value{
list of lists of ff vectors with genomic occuopancy
}
\description{
Individually converts a list of different sparse representation of dyad positions to nucleosome occupancy as list of ff vectors.
}
\author{
Mark Heron
}

