% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_functions.R
\name{cor_ff}
\alias{cor_ff}
\title{cor_ff}
\usage{
cor_ff(x, y, method = "pearson")
}
\arguments{
\item{x}{a ff vector}

\item{y}{a second ff vector}

\item{method}{("pearson") which correlation should be computed, either "pearson" or "spearman" (has to briefly transform the ff's to vectors for rank)}
}
\value{
pearson correlation between x and y
}
\description{
Computes the correlation between two vectors.
}
\author{
Mark Heron
}

