% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ff_functions.R
\name{cor_ff}
\alias{cor_ff}
\title{cor_ff}
\usage{
cor_ff(x, y)
}
\arguments{
\item{x}{a ff vector}

\item{y}{a second ff vector}
}
\value{
pearson correlation between x and y
}
\description{
Computes the correlation between two vectors.
}
\author{
Mark Heron
}

