% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_list_list_functions.R
\name{cor_ff_list_list}
\alias{cor_ff_list_list}
\title{cor_ff_list_list}
\usage{
cor_ff_list_list(data_list)
}
\arguments{
\item{data_list}{of the ff_list objects to compute the correlations between.}
}
\value{
correlation matrix
}
\description{
Computes the pairwise correlations between the list elements and saves them in a matrix
}
\author{
Mark Heron
}

