% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_functions.R
\name{cov_ff}
\alias{cov_ff}
\title{cov_ff}
\usage{
cov_ff(x, y, centered = FALSE)
}
\arguments{
\item{x}{(numeric ff vector)}

\item{y}{(numeric ff vector)}

\item{centered}{(boolean) are the means of x and y == 0?}
}
\value{
covariance between x and y
}
\description{
Computes the covariance between two ff vectors.
}

