% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cov_ff_list}
\alias{cov_ff_list}
\title{cov_ff_list}
\usage{
cov_ff_list(x, y, centered = FALSE)
}
\arguments{
\item{x}{a list of ff vectors}

\item{y}{a second list of ff vectors}

\item{centered}{boolean, is the mean of x and y already = 0 ?}
}
\value{
covariance between x and y
}
\description{
Computest the covariance between two ff vectors.
}
\author{
Mark Heron
}

