% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_list_functions.R
\name{cov_ff_list}
\alias{cov_ff_list}
\title{cov_ff_list}
\usage{
cov_ff_list(x, y, centered = FALSE)
}
\arguments{
\item{x}{a list of ff vectors}

\item{y}{a second list of ff vectors}

\item{centered}{boolean, is the mean of x and y already = 0 ?}
}
\value{
covariance between x and y
}
\description{
Computest the covariance between two ff vectors.
}
\author{
Mark Heron
}

