% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_list_functions.R
\name{df_list_to_ffdf_list}
\alias{df_list_to_ffdf_list}
\title{df_list_to_ffdf_list}
\usage{
df_list_to_ffdf_list(df_list)
}
\arguments{
\item{df_list}{to convert to a ffdf_list}
}
\value{
list of ffdf objects
}
\description{
Creates a copy of df_list where each element is saved as an ffdf object.
}
