% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_list_functions.R
\name{filter_column_range_ff_list}
\alias{filter_column_range_ff_list}
\title{filter_column_range_ff_list}
\usage{
filter_column_range_ff_list(ff_list, column, min, max)
}
\arguments{
\item{ff_list}{to filter}

\item{column}{based on which to filter}

\item{min}{allowed column value}

\item{max}{allowed column value}
}
\value{
list of ff's filtered by min/max on column
}
\description{
Filters the rows of the ff matrices in the list based on \code{column}, \code{min} and \code{max}.
}
