% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ff_list_functions.R
\name{hist_ff_list}
\alias{hist_ff_list}
\title{hist_ff_list}
\usage{
hist_ff_list(ff_list, column, ...)
}
\arguments{
\item{ff_list}{list with ff objects}

\item{column}{to plot the histogram for}

\item{...}{parameters that can be passed on the hist}
}
\description{
Plots a histogram for the values in column for every ff_list element.
}
\author{
Mark Heron
}

