% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{intra_scaling_list}
\alias{intra_scaling_list}
\title{intra_scaling_list}
\usage{
intra_scaling_list(list)
}
\arguments{
\item{list}{where each vector should be scaled individually}
}
\value{
intra scaled list
}
\description{
Scales each vector in the list to average 1.
Simple wrapper for lapply '/' mean
works on ff_lists
}
\author{
Mark Heron
}

