% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_list_list_functions.R
\name{likelihood_ff_list_list_vs_ff_list_list}
\alias{likelihood_ff_list_list_vs_ff_list_list}
\title{likelihood_ff_list_list_vs_ff_list_list}
\usage{
likelihood_ff_list_list_vs_ff_list_list(first_list, second_list)
}
\arguments{
\item{first_list}{list of the ff_list objects that represent the predictions}

\item{second_list}{list of the ff_list objects that represent the measurements}
}
\value{
likelihood matrix
}
\description{
Computes the pairwise likelihood between the list elements of two different lists and saves them in a matrix.
}

