% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ff_list_functions.R
\name{mae_ff_list}
\alias{mae_ff_list}
\title{mae_ff_list}
\usage{
mae_ff_list(predictions, measurements)
}
\arguments{
\item{predictions}{a list of ff vectors}

\item{measurements}{a second list of ff vectors}
}
\value{
mean absolute error
}
\description{
Computes the mean absolute error between two ff_lists.
}

