% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ff_list_functions.R
\name{make_prob_ff_list}
\alias{make_prob_ff_list}
\title{make_prob_ff_list}
\usage{
make_prob_ff_list(x)
}
\arguments{
\item{x}{list of ff_vectors}
}
\value{
probability version of x (sum = 1)
}
\description{
scales all list elements so their sum is 1, i.e. ig there are no negative values they can be used as a probability
}
\author{
Mark Heron
}

