% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_functions.R
\name{mean_list}
\alias{mean_list}
\title{mean_list}
\usage{
mean_list(x, na.rm = TRUE)
}
\arguments{
\item{x}{list of numeric vectors}

\item{na.rm}{should NA values be ignored (see sum)}
}
\value{
mean of the list
}
\description{
calculates the mean of all list elements together
}
\author{
Mark Heron
}

