% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{mean_list}
\alias{mean_list}
\title{mean_list}
\usage{
mean_list(x, na.rm = TRUE)
}
\arguments{
\item{x}{list of numeric vectors}

\item{na.rm}{should NA values be ignored (see \code{\link{sum}})}
}
\value{
(numeric) mean of all values in the list
}
\description{
Calculates the mean of all list elements together.
I.e. the mean of all values contained in list elements, \emph{not} the mean of the mean of all list elements!
}

