% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucular.R
\docType{package}
\name{nucular-package}
\alias{nucular-package}
\title{nucleosomal nucleotide frequencies Automatic Report}
\description{
Generates a set of figures for nucleosomal data, showing the fragment length distribution, occupancy, nucleotide frequencies around the mapped ends or dyad positions and correlations between different data sets
}
\details{
This package provides basic functions to handle nucleosome data, nuccontrol and nucppp actually do the reporting, so it might need a renaming.
}
\author{
Mark Heron
}

