% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{scale_list}
\alias{scale_list}
\title{scale_list}
\usage{
scale_list(x)
}
\arguments{
\item{x}{list of numeric vectors (or ff vectors)}
}
\value{
(list of numeric (ff) vectors) scaled \code{x} (i.e., mean_list == 1)
}
\description{
Scales all list elements by their combined mean.
Works with lists of ff vectors.
}

