% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_functions.R
\name{sd_ff}
\alias{sd_ff}
\title{sd_ff}
\usage{
sd_ff(x, centered = FALSE)
}
\arguments{
\item{x}{(numeric ff vector)}

\item{centered}{(boolean) is the mean of x == 0?}
}
\value{
standard deviation of x
}
\description{
Computes the standard deviation of a vector.
}

