% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ff_list_functions.R
\name{sd_ff_list}
\alias{sd_ff_list}
\title{sd_ff_list}
\usage{
sd_ff_list(x, centered = FALSE)
}
\arguments{
\item{x}{a list of ff vectors}

\item{centered}{boolean, is the mean already = 0 ?}
}
\value{
standard deviation of x
}
\description{
Computes the standard deviation of a vector.
}
\author{
Mark Heron
}

